+++ ci/deploy_benchmarks.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+set -e
+
+mkdir -p benchmarks/repository benchmarks/logs
+
+# Download them all
+aws s3 sync s3://nerabus/reco-sdaccel/benchmarks/ benchmarks/logs
+# Generate site
+docker run -v "$PWD":/mnt -w /mnt/benchmarks "parfunc/gipeda:latest" gipeda
+# Upload generated website
+aws s3 sync benchmarks/site s3://perf.reconfigureio-infra.com/reco-sdaccel/